requirements.txt
asgiref==3.5.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.0.4
django-ckeditor==6.4.0
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.1.0
python-dateutil==1.5
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
